Topic 62 - Factory Browser/ Factory Environment

+++ actions/commons/BaseTest.java
@@ -2,6 +2,8 @@ package commons;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
@@ -10,20 +12,33 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.edge.EdgeOptions;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.firefox.FirefoxOptions;
 import org.openqa.selenium.ie.InternetExplorerDriver;
 import org.openqa.selenium.logging.LogEntries;
 import org.openqa.selenium.logging.LogEntry;
 import org.openqa.selenium.opera.OperaDriver;
+import org.openqa.selenium.opera.OperaOptions;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.Assert;
 import org.testng.Reporter;
 import org.testng.annotations.BeforeSuite;
 
+import factoryEnvironment.Browser;
+import factoryEnvironment.BrowserstackFactory;
+import factoryEnvironment.CrossbrowserFactory;
+import factoryEnvironment.Environment;
+import factoryEnvironment.GridFactory;
+import factoryEnvironment.LambdaFactory;
+import factoryEnvironment.LocalFactory;
+import factoryEnvironment.SaucelabFactory;
 import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class BaseTest {
@@ -38,7 +53,42 @@ public class BaseTest {
 	protected BaseTest() {
 		log = LogFactory.getLog(getClass());
 	}
+	
+	protected WebDriver getBrowserDriver(String envName, String serverName, String  browserName, String ipAddress, String portNumber, String osName, String osVersion) {
+		switch (envName) {
+		case "local":
+			driver = new LocalFactory(browserName).createDriver();
+			break;
+		case "grid":
+			driver = new GridFactory(browserName, ipAddress, portNumber).createDriver();
+			break;
+		case "browserStack":
+			driver = new BrowserstackFactory(browserName, osName, osVersion).createDriver();
+			break;
+		case "saucelab":
+			driver = new SaucelabFactory(browserName, osName).createDriver();
+			break;
+		case "crossBrowser":
+			driver = new CrossbrowserFactory(browserName, osName).createDriver();
+			break;
+		case "lambda":
+			driver = new LambdaFactory(browserName, osName).createDriver();
+			break;
+		default:
+			driver = new LocalFactory(browserName).createDriver();
+			break;
+		}
+		
+		driver.manage().timeouts().implicitlyWait(GlobalConstants.LONG_TIMEOUT, TimeUnit.SECONDS);
+		driver.manage().window().maximize();
+		driver.get(getEnvironmentUrl(serverName));
+		return driver;
+	}
 
+	public WebDriver getDriverInstance() {
+		return this.driver;
+	}
+	
 	protected WebDriver getBrowserDriver(String browserName) {
 		if (browserName.equals("firefox")) {
 			// System.setProperty("webdriver.gecko.driver", projectPath +
@@ -96,45 +146,48 @@ public class BaseTest {
 		}
 
 		driver.manage().timeouts().implicitlyWait(GlobalConstants.LONG_TIMEOUT, TimeUnit.SECONDS);
+		driver.manage().window().maximize();
 		driver.get(GlobalConstants.PORTAL_DEV_URL);
 		return driver;
 	}
 
-	protected WebDriver getBrowserDriver2(String browserName) {
-		BrowserList browserList = BrowserList.valueOf(browserName.toUpperCase());
+	protected WebDriver getBrowserDriverWithEnvironmentName(String browserName, String environmentName) {
+		Browser browserList = Browser.valueOf(browserName.toUpperCase());
 
-		if (browserList == BrowserList.FIREFOX) {
+		if (browserList == Browser.FIREFOX) {
 			WebDriverManager.firefoxdriver().setup();
 			
 			System.setProperty(FirefoxDriver.SystemProperty.DRIVER_USE_MARIONETTE, "true");
 			System.setProperty(FirefoxDriver.SystemProperty.BROWSER_LOGFILE, GlobalConstants.PROJECT_PATH + "//browserLogs//FirefoxLog.log");
 			
+			FirefoxOptions options = new FirefoxOptions();
+			options.setAcceptInsecureCerts(false);
 			driver = new FirefoxDriver();
-		} else if (browserList == BrowserList.H_FIREFOX) {
+		} else if (browserList == Browser.H_FIREFOX) {
 			WebDriverManager.firefoxdriver().setup();
 			FirefoxOptions options = new FirefoxOptions();
 			options.addArguments("--headless");
 			options.addArguments("window-size=1920x1080");
 			driver = new FirefoxDriver(options);
-		} else if (browserList == BrowserList.CHROME) {
+		} else if (browserList == Browser.CHROME) {
 			WebDriverManager.chromedriver().setup();
 			driver = new ChromeDriver();
-		} else if (browserList == BrowserList.H_CHROME) {
+		} else if (browserList == Browser.H_CHROME) {
 			WebDriverManager.chromedriver().setup();
 			ChromeOptions options = new ChromeOptions();
 			options.addArguments("--headless");
 			options.addArguments("window-size=1920x1080");
 			driver = new ChromeDriver(options);
-		} else if (browserList == BrowserList.EDGE) {
+		} else if (browserList == Browser.EDGE) {
 			WebDriverManager.edgedriver().setup();
 			driver = new EdgeDriver();
-		} else if (browserList == BrowserList.IE) {
+		} else if (browserList == Browser.IE) {
 			WebDriverManager.iedriver().arch32().setup();
 			driver = new InternetExplorerDriver();
-		} else if (browserList == BrowserList.OPERA) {
+		} else if (browserList == Browser.OPERA) {
 			WebDriverManager.operadriver().setup();
 			driver = new OperaDriver();
-		} else if (browserList == BrowserList.COCCOC) {
+		} else if (browserList == Browser.COC_COC) {
 			// Cốc Cốc brwoser trừ đi 5-6 version ra chromedriver
 			WebDriverManager.chromedriver().driverVersion("93.0.4577.63").setup();
 			ChromeOptions options = new ChromeOptions();
@@ -145,7 +198,7 @@ public class BaseTest {
 			}
 			driver = new ChromeDriver(options);
 
-		} else if (browserList == BrowserList.BRAVE) {
+		} else if (browserList == Browser.BRAVE) {
 			// Brave browser version nào dùng chromedriver version đó
 			WebDriverManager.chromedriver().driverVersion("95.0.4638.17").setup();
 			ChromeOptions options = new ChromeOptions();
@@ -154,8 +207,11 @@ public class BaseTest {
 		} else {
 			// throw new BrowserNotSupport(browserName);
 		}
+		
 		driver.manage().timeouts().implicitlyWait(GlobalConstants.LONG_TIMEOUT, TimeUnit.SECONDS);
-		driver.get(GlobalConstants.PORTAL_DEV_URL);
+		driver.manage().window().maximize();
+		//driver.get(getEnvironmentUrl(environmentName));
+		driver.get(getEnvironmentUrl(environmentName));
 		return driver;
 	}
 
@@ -215,22 +271,94 @@ public class BaseTest {
 		return driver;
 	}
 
-	public WebDriver getDriverInstance() {
-		return this.driver;
+	protected WebDriver getBrowserDriverGrid(String browserName, String appUrl, String osName, String ipAddress, String portNumber) {
+		DesiredCapabilities capability = null;
+		Platform platform = null;
+
+		if (osName.contains("windows")) {
+			platform = Platform.WINDOWS;
+		} else {
+			platform = Platform.MAC;
+		}
+
+		switch (browserName) {
+			case "firefox" :
+				capability = DesiredCapabilities.firefox();
+				capability.setBrowserName("firefox");
+				capability.setPlatform(platform);
+
+				FirefoxOptions fOptions = new FirefoxOptions();
+				fOptions.merge(capability);
+				break;
+			case "chrome" :
+				capability = DesiredCapabilities.chrome();
+				capability.setBrowserName("chrome");
+				capability.setPlatform(platform);
+
+				ChromeOptions cOptions = new ChromeOptions();
+				cOptions.merge(capability);
+				break;
+			case "edge" :
+				capability = DesiredCapabilities.edge();
+				capability.setBrowserName("edge");
+				capability.setPlatform(platform);
+
+				EdgeOptions eOptions = new EdgeOptions();
+				eOptions.merge(capability);
+				break;
+			case "opera" :
+				capability = DesiredCapabilities.opera(); //kệ nó nếu nó bị deprecated
+				capability.setBrowserName("opera");
+				capability.setPlatform(platform);
+
+				OperaOptions oOptions = new OperaOptions();
+				oOptions.merge(capability);
+				break;
+			default :
+				throw new RuntimeException("Browser is not valid!");
+		}
+
+		try {
+			driver = new RemoteWebDriver(new URL(String.format("http://%s:%s/wd/hub", ipAddress, portNumber)), capability);
+		} catch (MalformedURLException e) {
+			e.printStackTrace();
+		}
+		
+		driver.manage().timeouts().implicitlyWait(GlobalConstants.LONG_TIMEOUT, TimeUnit.SECONDS);
+		driver.manage().window().maximize();
+		driver.get(appUrl);
+		return driver;
+
 	}
 
-	/*
-	 * protected String getEnvironmentUrl(String serverName) { String envUrl = null;
-	 * EnvironmentList environment =
-	 * EnvironmentList.valueOf(serverName.toUpperCase()); if (environment ==
-	 * EnvironmentList.DEV) { envUrl = "https://demo.nopcommerce.com/"; } else if
-	 * (environment == EnvironmentList.TESTING) { envurl =
-	 * "https://admin-demo.nopcommerce.com"; } else if (environment ==
-	 * EnvironmentList.STAGING) { envurl = "https://staging.orangehrmlive.com";| }
-	 * else if (environment == EnvironmentList. PRODUCTION) { envUrl =
-	 * "https://production.orangehrmlive.com"; } return envUrl; }
-	 */
+	protected String getEnvironmentUrl(String serverName) {
+		String envUrl = null;
+		Environment environment = Environment.valueOf(serverName.toUpperCase());
 
+		switch (environment) {
+		case DEV:
+			envUrl = "https://demo.nopcommerce.com/";
+			break;
+		case TESTING:
+			envUrl = "https://testing.nopcommerce.com/";
+			break;
+		case STAGING:
+			envUrl = "https://staging.nopcommerce.com/";
+			break;
+		case PRE_PROD:
+			envUrl = "https://pre-prod.nopcommerce.com/";
+			break;
+		case PROD:
+			envUrl = "https://prod.nopcommerce.com/";
+			break;
+		default:
+			envUrl = null;
+			break;
+		}
+		
+		return envUrl;
+	}
+	 
 	protected boolean verifyTrue(boolean condition) {
 		boolean pass = true;
 		try {
@@ -299,51 +427,53 @@ public class BaseTest {
 		}
 	}
 
-	protected void closeBrowserDriver() {
-		String cmd = null;
-		try {
-			String osName = System.getProperty("os.name").toLowerCase();
-			log.info("OS name = " + osName);
-
-			String driverInstanceName = driver.toString().toLowerCase();
-			log.info("Driver instance name = " + driverInstanceName);
-
-			String browserDriverName = null;
-
-			if (driverInstanceName.contains("chrome")) {
-				browserDriverName = "chromedriver";
-			} else if (driverInstanceName.contains("internetexplorer")) {
-				browserDriverName = "IEDriverServer";
-			} else if (driverInstanceName.contains("firefox")) {
-				browserDriverName = "geckodriver";
-			} else if (driverInstanceName.contains("edge")) {
-				browserDriverName = "msedgedriver";
-			} else if (driverInstanceName.contains("opera")) {
-				browserDriverName = "operadriver";
-			} else {
-				browserDriverName = "safaridriver";
-			}
-
-			if (osName.contains("window")) {
-				cmd = "taskkill /F /FI \"IMAGENAME eq " + browserDriverName + "*\"";
-			} else {
-				cmd = "pkill " + browserDriverName;
-			}
-
-			if (driver != null) {
-				driver.manage().deleteAllCookies();
-				driver.quit();
-			}
-		} catch (Exception e) {
-			log.info(e.getMessage());
-		} finally {
+	protected void closeBrowserDriver(String envName) {
+		if(envName.equals("local") || envName.equals("grid")) {
+			String cmd = null;
 			try {
-				Process process = Runtime.getRuntime().exec(cmd);
-				process.waitFor();
-			} catch (IOException e) {
-				e.printStackTrace();
-			} catch (InterruptedException e) {
-				e.printStackTrace();
+				String osName = System.getProperty("os.name").toLowerCase();
+				log.info("OS name = " + osName);
+
+				String driverInstanceName = driver.toString().toLowerCase();
+				log.info("Driver instance name = " + driverInstanceName);
+
+				String browserDriverName = null;
+
+				if (driverInstanceName.contains("chrome")) {
+					browserDriverName = "chromedriver";
+				} else if (driverInstanceName.contains("internetexplorer")) {
+					browserDriverName = "IEDriverServer";
+				} else if (driverInstanceName.contains("firefox")) {
+					browserDriverName = "geckodriver";
+				} else if (driverInstanceName.contains("edge")) {
+					browserDriverName = "msedgedriver";
+				} else if (driverInstanceName.contains("opera")) {
+					browserDriverName = "operadriver";
+				} else {
+					browserDriverName = "safaridriver";
+				}
+
+				if (osName.contains("window")) {
+					cmd = "taskkill /F /FI \"IMAGENAME eq " + browserDriverName + "*\"";
+				} else {
+					cmd = "pkill " + browserDriverName;
+				}
+
+				if (driver != null) {
+					driver.manage().deleteAllCookies();
+					driver.quit();
+				}
+			} catch (Exception e) {
+				log.info(e.getMessage());
+			} finally {
+				try {
+					Process process = Runtime.getRuntime().exec(cmd);
+					process.waitFor();
+				} catch (IOException e) {
+					e.printStackTrace();
+				} catch (InterruptedException e) {
+					e.printStackTrace();
+				}
 			}
 		}
 	}
